requirements.txt
pandas
numpy==1.26.4
matplotlib
altair==4.2.0
scikit-learn
pyyaml
joblib
seaborn
tqdm
openpyxl
transformers
torch
numba
combo
pyod
shap
rrcf
streamlit